move components around
