+++ client/src/App.tsx
@@ -1,24 +1,18 @@
 import * as React from "react";
 import { AuthProvider } from "./auth";
 import { ReactQueryDevtools } from "react-query-devtools";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 import { QueryConfigProvider, useActivitiesPrefetch } from "./data";
-import { Navigation, PageNotFound, UIProvider } from "./components";
-import { Welcome, Profile, Charts, ActivityList } from "./pages";
+import { AppContainer, Navigation, UIProvider } from "./components";
+import { Pages } from "./pages";
 
 function App() {
   useActivitiesPrefetch();
   return (
-    <BrowserRouter>
+    <AppContainer>
       <Navigation />
-      <Switch>
-        <Route exact path="/" component={Welcome} />
-        <Route exact path="/profile" component={Profile} />
-        <Route exact path="/charts" component={Charts} />
-        <Route exact path="/activity-list" component={ActivityList} />
-        <Route path="/" component={PageNotFound} />
-      </Switch>
-    </BrowserRouter>
+      <Pages />
+    </AppContainer>
   );
 }
 
@@ -27,8 +21,10 @@ function AppWithProviders() {
     <AuthProvider>
       <QueryConfigProvider>
         <UIProvider>
-          <App />
-          <ReactQueryDevtools initialIsOpen={false} />
+          <BrowserRouter>
+            <App />
+            <ReactQueryDevtools initialIsOpen={false} />
+          </BrowserRouter>
         </UIProvider>
       </QueryConfigProvider>
     </AuthProvider>
