+++ client/src/components/index.ts
@@ -3,7 +3,6 @@ export { DateFilterForm } from "./forms/DateFilterForm";
 export { FileUploadForm } from "./forms/FileUploadForm";
 
 export { Navigation } from "./navigation/Navigation";
-export { PageNotFound } from "./navigation/PageNotFound";
 
 export { ErrorView } from "./states/ErrorView";
 export { Loading } from "./states/Loading";
@@ -14,5 +13,6 @@ export { BarChart } from "./visualization/Charts/BarChart";
 export { SummaryBarChart } from "./visualization/SummaryCharts/SummaryBarChart";
 export { SummaryPieChart } from "./visualization/SummaryCharts/SummaryPieChart";
 
-export { ModalDialog } from "./material-ui/ModalDialog";
-export { UIProvider } from "./material-ui/UIProvider";
+export { ModalDialog } from "./ModalDialog";
+export { UIProvider } from "./UIProvider";
+export { AppContainer } from "./AppContainer";
