+++ client/src/pages/index.ts
@@ -1,4 +1,1 @@
-export { Welcome } from "./Welcome";
-export { Charts } from "./Charts";
-export { Profile } from "./Profile";
-export { ActivityList } from "./ActivityList";
+export { Pages } from "./Pages";
